+++ src/features/schedule/lib/use-address-modal.ts
@@ -1,0 +1,24 @@
+import { useEffect, useState } from "react";
+
+export const useAddressModal = () => {
+  const [isVisible, setIsVisible] = useState(false);
+
+  const open = () => setIsVisible(true);
+  const close = () => setIsVisible(false);
+
+  // ESC 키로 닫기
+  useEffect(() => {
+    const handleEscape = (e: KeyboardEvent) => {
+      if (e.key === "Escape" && isVisible) {
+        close();
+      }
+    };
+
+    if (isVisible) {
+      document.addEventListener("keydown", handleEscape);
+      return () => document.removeEventListener("keydown", handleEscape);
+    }
+  }, [isVisible]);
+
+  return { isVisible, open, close };
+};
